backend/requirements.txt
fastapi
uvicorn
httpx
pytesseract
pillow
python-dotenv
pydantic-settings
python-multipart
